Modified pwi-example to allow AFE register access

+++ arrus/core/devices/us4r/Us4RImpl.cpp
@@ -277,11 +277,11 @@ void Us4RImpl::setActiveTermination(std::optional<uint16> value) {
 }
 
 uint8_t Us4RImpl::getNumberOfUs4OEMs() {
-    return us4oems.size();
+    return (uint8_t)us4oems.size();
 }
 
 float Us4RImpl::getSamplingFrequency() const {
-    return us4oems[0]->getSamplingFrequency();
+    return (float)us4oems[0]->getSamplingFrequency();
 }
 
 }
